Lyricist/requirements.txt
discord.py==1.7.3
PyNaCl==1.4.0
async-timeout==3.0.1
aiohttp==3.7.4.post0
beautifulsoup4==4.9.3
config==0.5.0.post0
lyricsgenius==3.0.1
python-dotenv==0.18.0
discord==1.7.3